requirements.txt
pyspark==3.5.1
delta-spark==3.2.0
PyYAML==6.0.2
rapidfuzz==3.9.6
duckdb==1.0.0
pandas==2.2.2
pytest==8.3.2
